Restructure project into core modules
