+++ src/components/HudBar.vue
@@ -1,0 +1,9 @@
+<template>
+  <div class="hud-bar"></div>
+</template>
+
+<script lang="ts">
+export default {
+  name: 'HudBar'
+}
+</script>
