+++ src/components/MiniMap.vue
@@ -1,0 +1,9 @@
+<template>
+  <div class="mini-map"></div>
+</template>
+
+<script lang="ts">
+export default {
+  name: 'MiniMap'
+}
+</script>
