+++ src/components/TowerPanel.vue
@@ -1,0 +1,9 @@
+<template>
+  <div class="tower-panel"></div>
+</template>
+
+<script lang="ts">
+export default {
+  name: 'TowerPanel'
+}
+</script>
