+++ src/core/engine/Game.ts
@@ -1,0 +1,4 @@
+export class Game {
+  start() {}
+  update(_dt: number) {}
+}
