+++ src/core/engine/Input.ts
@@ -1,0 +1,3 @@
+export class Input {
+  onPointer(_cb: (x: number, y: number) => void) {}
+}
