+++ src/core/engine/Renderer.ts
@@ -1,0 +1,3 @@
+export class Renderer {
+  init(_canvas: HTMLCanvasElement) {}
+}
