+++ src/core/engine/Time.ts
@@ -1,0 +1,5 @@
+export class Time {
+  dt = 0
+  speed = 1
+  paused = false
+}
