+++ src/core/gameplay/DamageSystem.ts
