+++ src/core/gameplay/EnemyManager.ts
@@ -1,0 +1,8 @@
+import { Enemy } from './Enemy'
+
+export class EnemyManager {
+  enemies: Enemy[] = []
+  update(dt: number) {
+    for (const e of this.enemies) e.update(dt)
+  }
+}
