+++ src/core/gameplay/LevelManager.ts
@@ -1,0 +1,8 @@
+import type { LevelConfig } from '../data/types'
+
+export class LevelManager {
+  current?: LevelConfig
+  load(level: LevelConfig) {
+    this.current = level
+  }
+}
