+++ src/core/gameplay/Projectile.ts
