+++ src/core/gameplay/ProjectileManager.ts
@@ -1,0 +1,5 @@
+import type { Projectile } from './Projectile'
+
+export class ProjectileManager {
+  projectiles: Projectile[] = []
+}
