+++ src/core/gameplay/StatusSystem.ts
@@ -1,0 +1,3 @@
+export class StatusSystem {
+  update(_dt: number) {}
+}
