+++ src/core/gameplay/Targeting.ts
@@ -1,0 +1,5 @@
+import { Enemy } from './Enemy'
+
+export function findTarget(enemies: Enemy[]): Enemy | undefined {
+  return enemies[0]
+}
