+++ src/core/gameplay/TowerManager.ts
@@ -1,0 +1,5 @@
+import { Tower } from './Tower'
+
+export class TowerManager {
+  towers: Tower[] = []
+}
