+++ src/core/gameplay/WaveManager.ts
@@ -1,0 +1,6 @@
+import type { Wave } from '../data/types'
+
+export class WaveManager {
+  waves: Wave[] = []
+  start() {}
+}
