+++ src/core/utils/Pool.ts
@@ -1,0 +1,11 @@
+export class Pool<T> {
+  private items: T[] = []
+
+  acquire(create: () => T): T {
+    return this.items.pop() ?? create()
+  }
+
+  release(item: T) {
+    this.items.push(item)
+  }
+}
