+++ src/core/utils/RNG.ts
@@ -1,0 +1,10 @@
+export class RNG {
+  constructor(private seed = Date.now()) {}
+
+  next() {
+    let t = this.seed += 0x6D2B79F5
+    t = Math.imul(t ^ (t >>> 15), t | 1)
+    t ^= t + Math.imul(t ^ (t >>> 7), t | 61)
+    return ((t ^ (t >>> 14)) >>> 0) / 4294967296
+  }
+}
