+++ src/router/index.js
@@ -1,12 +1,10 @@
 import { createRouter, createWebHistory } from 'vue-router'
 import HomeView from '../views/HomeView.vue'
 import GameView from '../views/GameView.vue'
-import TowerDefenseView from '../views/TowerDefenseView.vue'
 
 const routes = [
   { path: '/', name: 'home', component: HomeView },
-  { path: '/game', name: 'game', component: GameView },
-  { path: '/tower', name: 'tower', component: TowerDefenseView }
+  { path: '/game', name: 'game', component: GameView }
 ]
 
 const router = createRouter({
