--- src/views/TowerDefenseView.vue
@@ -1,13 +1,0 @@
-<template>
-  <TowerDefense />
-</template>
-
-<script lang="ts">
-import { defineComponent } from 'vue'
-import TowerDefense from '../components/TowerDefense.vue'
-
-export default defineComponent({
-  name: 'TowerDefenseView',
-  components: { TowerDefense }
-})
-</script>
